feat: generalized sample loading
