--- Editor/Utils/QuickStartHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a18cc655bd80acf4db55549f1dd30bd1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
